fix type
git-svn-id: svn+ssh://svn.savannah.nongnu.org/gap/trunk@3144 d25b4930-3292-e311-b074-0010c6de5da8

+++ user-apps/LaternaMagica/AppController.m
@@ -423,7 +423,7 @@
 	{
 	  NSWorkspace *ws;
 	  NSString *fileOperation;
-	  int opTag;
+	  NSInteger opTag;
 	  NSString *folder;
 	  NSString *file;
